refactor: rename `run workflow` command to `run-workflow`
This is to make way for the new `garden run` command.

BREAKING CHANGE:

The `garden run workflow` command has been renamed to `garden run-workflow`
(note the dash).
